+++ tests/sfc.R
@@ -100,6 +100,7 @@ ls - 2
 str(x)
 nc = st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
 str(nc)
+st_as_sfc(st_bbox(nc))
 
 st_agr("constant")
 st_agr()