+++ R/tidyverse.R
@@ -13,7 +13,14 @@
 #' nc %>% filter(AREA > .1) %>% plot()
 #' @export
 filter_.sf <- function(.data, ..., .dots) {
-	st_as_sf(NextMethod())
+	#st_as_sf(NextMethod())
+	sf_column = attr(.data, "sf_column")
+	geom = .data[[sf_column]]
+	.data[[sf_column]] = 1:nrow(.data)
+	ret = NextMethod()
+	sel = ret[[sf_column]]
+	ret[[sf_column]] = geom[sel]
+	st_as_sf(ret)
 }
 #' @name dplyr
 #' @export